use uintptr_t instead of unsigned long - fixes build error on win64

+++ include/rtl_433_devices.h
@@ -63,7 +63,7 @@ typedef struct {
 	float reset_limit;
 	int (*json_callback)(bitbuffer_t *bitbuffer);
 	unsigned int disabled;
-	unsigned long demod_arg;	// Decoder specific optional argument (may be pointer to struct)
+	uintptr_t demod_arg;	// Decoder specific optional argument (may be pointer to struct)
 	char **fields;			// List of fields this decoder produces; required for CSV output. NULL-terminated.
 } r_device;
